Fixed issue with flag, lat and lon fields, added upload button and preview of uploaded data
The query didn't filter by flag since last commit (`81fa9c7`) because the name of the variable was wrong. Lat and Lon fields were filtered by the actual inputs, but the data in the table is binned to 100th or 10th of degree, so the input needs to be multiplied by 100 and 10 respectively to be able to query. The upload button and data visualizer work and have been styled with CSS, but it might be good to find a way to just use the queried data for the analysis without needing to download and upload. Some reactive functions have also been deleted because useless. Date of copyrighted material has been updated.