Debug, right sidebar into footer, summary preview
Now uploading a csv file does not freeze the query option. This happened because I assigned two different functions to the same name my_data, instead of assigning one function with if conditions nested within (as it is now).

To give more space to the tables, I turned the right sidebar into a footer. It is a bit empty for now, but at least it does not occupy useful space. This called for a banner instead of a logo for fishRman.

Summaries now have a preview.