clipping data to area now affects descriptive analysis
Summaries now change according to the clipped data, and can be performed on those. Loading a new csv or querying new data will uncheck the clip data box and overwrite this step.

Although the result for this particular issue is achieved, I am not satisfied with the overall quality of the code. A new issue on code keeping will be raised.