code keeping, debugged "upload gpkg", more in comment
The boundaries of the plot are now in a function of its own, and can now be inputed by the user, for a "zoom in/zoom out" effect, to give some reference to smaller areas far from shore.

Resolution of the plot is now a function of its own.

"Upload gpkg" now checks whether the layer "GFW" exists, before trying to st_read the file, thus avoiding an error.

my_data is now also created by uploading a gpkg