simplified UI/UX and reactive CSS
The types of data are no longer divided in spatial data and normal dataframe. We always use sf objects now (spatial data), and we derive the dataframe in a few moments with as.data.frame() and then removing the "geometry" column. This makes the "Convert" button and various tabs useless, and so they have been deleted. More commentary to the code will be added soon. The new UI is now optimised for various screen sizes, including smaller ones such as smartphones.