remade fishRman as golem package
Since the changes were substantial, I have waited until the app was at least almost fully functioning. Now it makes calls to a still non-published API and requires files that were too big to fit in a package, but it works when the GFW csv files are uploaded with the specific button. Soon, I will take care of these other points and write tests for CI/CD.